Fix format warnings

+++ iked/util.c
@@ -425,10 +425,10 @@ print_spi(uint64_t spi, int size)
 		snprintf(ptr, 32, "0x%08x", (uint32_t)spi);
 		break;
 	case 8:
-		snprintf(ptr, 32, "0x%016llx", spi);
+		snprintf(ptr, 32, "0x%016llx", (unsigned long long)spi);
 		break;
 	default:
-		snprintf(ptr, 32, "%llu", spi);
+		snprintf(ptr, 32, "%llu", (unsigned long long)spi);
 		break;
 	}
